refactor: move all the marketplace related code to its own package
Instead of having everything scattered around, it makes sense having
everything in one place, to reduce coupling code to specific
implementations.
